refactor: moved gauges.py into src/ddr
