+++ src/ddr/__init__.py
@@ -1,6 +1,7 @@
 from ddr.analysis.metrics import Metrics
+from ddr.dataset.Gauges import Gauge, GaugeSet, validate_gages
 from ddr.dataset.streamflow import StreamflowReader
 from ddr.nn.kan import kan
 from ddr.routing.dmc import dmc
 
-__all__ = ["dmc", "kan", "StreamflowReader", "Metrics"]
+__all__ = ["dmc", "kan", "StreamflowReader", "Metrics", "Gauge", "GaugeSet", "validate_gages"]
